added BaseUserRole to make it inheritable
